backend/cloudFunctions/getTeamMembers/requirements.txt
google-cloud-firestore
firebase-admin
google-auth
flask
google-api-python-client
flask
functions-framework